agent_coordination/publish_analytics_completion.py
#!/usr/bin/env python3
"""Publish analytics task completion message."""

import asyncio
from nats.aio.client import Client as NATS
from datetime import datetime


async def publish_completion():
    """Publish analytics completion message to NATS."""
    nc = NATS()

    try:
        await nc.connect("nats://localhost:4222")
        print("✓ Connected to NATS server")

        message = {
            "timestamp": datetime.utcnow().isoformat(),
            "agent_id": "Agent-Analytics",
            "task": "Phase 2.5: Analytics Setup",
            "status": "RESEARCH COMPLETE",
            "summary": {
                "completed": [
                    "Researched Ghost built-in analytics",
                    "Compared privacy-focused alternatives (Plausible, Fathom, Simple Analytics)",
                    "Recommendation: Use Ghost built-in analytics (included, GDPR-compliant, zero setup)",
                    "Created comprehensive setup guide",
                    "Created implementation checklist"
                ],
                "documents": [
                    "/plans/analytics-setup-research.md",
                    "/plans/analytics-implementation-checklist.md"
                ],
                "recommendation": "Ghost built-in analytics",
                "implementation_time": "~1.5 hours",
                "ready_for": "Implementation by user or browser agent"
            }
        }

        await nc.publish(
            "coordination",
            str(message).encode()
        )
        print("✓ Analytics completion message published")

        await nc.flush()
        await nc.close()

        print("\n✨ Success! View at: http://localhost:8001")

    except Exception as e:
        print(f"❌ Failed to connect to NATS: {e}")
        print("Note: NATS server may not be running. This is optional.")


if __name__ == "__main__":
    asyncio.run(publish_completion())
