agent_coordination/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
nats-py>=2.7.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
